# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API Gestión de Solicitudes Académicas
  version: 2.1.0
  description: |
    API REST para sistema de triage y gestión de solicitudes académicas.

    Funcionalidades principales:
    - Registro de solicitudes académicas
    - Consulta y seguimiento del estado de solicitudes
    - Clasificación de solicitudes
    - Priorización de solicitudes
    - Asignación de responsables
    - Gestión controlada del ciclo de vida
    - Cierre formal de solicitudes
    - Historial auditable
    - Sugerencias asistidas por motor de reglas / IA opcional
    - Autenticación y autorización con JWT
servers:
  - url: http://localhost:8080
    description: Servidor local
...
security:
  - bearerAuth: []
paths:
...
  /api/usuarios/responsables-activos:
    get:
      tags: [Usuarios]
      summary: Listar responsables activos
      description: |
        Lista usuarios activos elegibles para asignación de solicitudes.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.
      responses:
        '200':
          description: Responsables activos obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUsuarioSimpleList'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes:
    post:
      tags: [Solicitudes]
      summary: Registrar una solicitud académica
      description: |
        Registra una nueva solicitud académica.
        La identidad del solicitante se obtiene desde el usuario autenticado.
        La solicitud queda inicialmente en estado `REGISTRADA`.
        **Roles permitidos:** ESTUDIANTE, ADMINISTRATIVO, COORDINADOR.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitudCreate'
      responses:
        '201':
          description: Solicitud registrada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Error de validación o regla de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    get:
      tags: [Solicitudes]
      summary: Consultar solicitudes por filtros
      description: |
        Consulta solicitudes con filtros opcionales y paginación.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.
      parameters:
        - in: query
          name: estado
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - in: query
          name: tipo
          schema:
            $ref: '#/components/schemas/TipoSolicitud'
        - in: query
          name: prioridad
          schema:
            $ref: '#/components/schemas/Prioridad'
        - in: query
          name: responsableId
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 2
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            example: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 10
        - in: query
          name: sort
          schema:
            type: string
            example: fechaHoraRegistro,desc
      responses:
        '200':
          description: Consulta realizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudPage'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}:
    get:
      tags: [Solicitudes]
      summary: Obtener solicitud por ID
      description: |
        Obtiene el detalle de una solicitud académica.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/clasificar:
    put:
      tags: [Solicitudes]
      summary: Clasificar solicitud
      description: |
        Define o ajusta el tipo de solicitud.
        Si la solicitud está en estado `REGISTRADA`, esta operación la mueve a `CLASIFICADA`.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - No se permite modificar una solicitud cerrada
        - Debe quedar registro en historial
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudRequest'
      responses:
        '200':
          description: Solicitud clasificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Datos inválidos o regla de negocio incumplida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflicto de estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/priorizar:
    put:
      tags: [Solicitudes]
      summary: Priorizar solicitud
      description: |
        Define o ajusta la prioridad de atención de una solicitud.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - No se permite modificar una solicitud cerrada
        - Idealmente la solicitud ya debe estar clasificada
        - Debe quedar registro en historial
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriorizarSolicitudRequest'
      responses:
        '200':
          description: Solicitud priorizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Datos inválidos o regla de negocio incumplida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflicto de estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/asignar:
    put:
      tags: [Solicitudes]
      summary: Asignar responsable
      description: |
        Asigna un responsable activo a la solicitud.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - No se permite modificar una solicitud cerrada
        - El responsable debe estar activo
        - Se registra evento en historial
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableRequest'
      responses:
        '200':
          description: Responsable asignado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Responsable inválido o inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud o responsable no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflicto de estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/iniciar-atencion:
    put:
      tags: [Solicitudes]
      summary: Iniciar atención de la solicitud
      description: |
        Cambia el estado de la solicitud a `EN_ATENCION`.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - Solo permitido desde estado `CLASIFICADA`
        - No se permite sobre solicitudes cerradas
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservacionRequest'
      responses:
        '200':
          description: Solicitud pasada a EN_ATENCION correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Transición inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/marcar-atendida:
    put:
      tags: [Solicitudes]
      summary: Marcar solicitud como atendida
      description: |
        Cambia el estado de la solicitud a `ATENDIDA`.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - Solo permitido desde estado `EN_ATENCION`
        - No se permite sobre solicitudes cerradas
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservacionRequest'
      responses:
        '200':
          description: Solicitud marcada como atendida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Transición inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/cerrar:
    put:
      tags: [Solicitudes]
      summary: Cerrar solicitud
      description: |
        Cierra formalmente la solicitud registrando observación de cierre.
        Solo se permite si la solicitud está en estado `ATENDIDA`.

        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.

        Restricciones:
        - Una solicitud cerrada no puede modificarse posteriormente
        - El cierre genera historial auditable
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudRequest'
      responses:
        '200':
          description: Solicitud cerrada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '400':
          description: Datos inválidos o regla de negocio incumplida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: No puede cerrarse en el estado actual o ya está cerrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/solicitudes/{id}/historial:
    get:
      tags: [Historial]
      summary: Obtener historial de la solicitud
      description: |
        Devuelve el historial auditable de acciones de la solicitud.
        **Roles permitidos:** ADMINISTRATIVO, COORDINADOR.
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      responses:
        '200':
          description: Historial obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHistorialList'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/mis-solicitudes:
    get:
      tags: [Mis Solicitudes]
      summary: Listar solicitudes del usuario autenticado
      description: |
        Retorna las solicitudes creadas por el usuario autenticado.
        **Roles permitidos:** ESTUDIANTE, ADMINISTRATIVO, COORDINADOR.
      parameters:
        - in: query
          name: estado
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            example: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 10
        - in: query
          name: sort
          schema:
            type: string
            example: fechaHoraRegistro,desc
      responses:
        '200':
          description: Solicitudes obtenidas correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudPage'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/mis-solicitudes/{id}:
    get:
      tags: [Mis Solicitudes]
      summary: Obtener detalle de una solicitud propia
      description: |
        Retorna el detalle de una solicitud perteneciente al usuario autenticado.
        **Roles permitidos:** ESTUDIANTE, ADMINISTRATIVO, COORDINADOR.
      parameters:
        - $ref: '#/components/parameters/SolicitudId'
      responses:
        '200':
          description: Solicitud obtenida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSolicitudResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: No autorizado para consultar esa solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SolicitudId:
      name: id
      in: path
      required: true
      description: ID de la solicitud
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 1
  schemas:
    TipoSolicitud:
      type: string
      enum:
        - REGISTRO_ASIGNATURAS
        - HOMOLOGACION
        - CANCELACION_ASIGNATURAS
        - SOLICITUD_CUPOS
        - CONSULTA_ACADEMICA
        - OTRO
      example: HOMOLOGACION
    Prioridad:
      type: string
      enum: [BAJA, MEDIA, ALTA, CRITICA]
      example: ALTA
    EstadoSolicitud:
      type: string
      enum: [REGISTRADA, CLASIFICADA, EN_ATENCION, ATENDIDA, CERRADA]
      example: REGISTRADA
    CanalOrigen:
      type: string
      enum: [CSU, CORREO, SAC, TELEFONICO, PRESENCIAL, OTRO]
      example: CORREO
    ImpactoAcademico:
      type: string
      enum: [BAJO, MEDIO, ALTO, CRITICO]
      example: ALTO
    RolUsuario:
      type: string
      enum: [ESTUDIANTE, ADMINISTRATIVO, COORDINADOR]
      example: ESTUDIANTE
...
    ApiResponseBase:
      type: object
      required: [success, message, timestamp]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operación exitosa
        timestamp:
          type: string
          format: date-time
          example: "2026-03-14T15:30:00"
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-03-14T15:30:00"
        status:
          type: integer
          example: 409
        error:
          type: string
          example: Conflict
        message:
          type: string
          example: La solicitud ya se encuentra cerrada y no puede modificarse
        path:
          type: string
          example: /api/solicitudes/1/cerrar
...
    ApiResponseUsuarioSimpleList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UsuarioSimple'
    ApiResponseSolicitudResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/SolicitudResponse'
    ApiResponseSolicitudPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PageSolicitudResponse'
    ApiResponseHistorialList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/HistorialResponse'
...
    UsuarioSimple:
      type: object
      required: [id, username, nombreCompleto, rol, activo]
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: coordinador1
        nombreCompleto:
          type: string
          example: Coordinador Académico
        rol:
          $ref: '#/components/schemas/RolUsuario'
        activo:
          type: boolean
          example: true
...
    SolicitudCreate:
      type: object
      required: [descripcion, canalOrigen]
      properties:
        descripcion:
          type: string
          minLength: 10
          maxLength: 2000
          example: Solicito homologación de Matemáticas I cursada en otra universidad.
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        impactoAcademico:
          allOf:
            - $ref: '#/components/schemas/ImpactoAcademico'
          nullable: true
        fechaLimite:
          type: string
          format: date
          nullable: true
          example: 2026-03-20
    ClasificarSolicitudRequest:
      type: object
      required: [tipoSolicitud]
      properties:
        tipoSolicitud:
          $ref: '#/components/schemas/TipoSolicitud'
        observacion:
          type: string
          minLength: 3
          maxLength: 500
          nullable: true
          example: Se clasifica como homologación por el contenido de la solicitud.
    PriorizarSolicitudRequest:
      type: object
      required: [prioridad, justificacionPrioridad]
      properties:
        prioridad:
          $ref: '#/components/schemas/Prioridad'
        justificacionPrioridad:
          type: string
          minLength: 5
          maxLength: 500
          example: Existe cercanía con fecha de matrícula y alto impacto académico.
    AsignarResponsableRequest:
      type: object
      required: [responsableId]
      properties:
        responsableId:
          type: integer
          format: int64
          minimum: 1
          example: 2
    ObservacionRequest:
      type: object
      properties:
        observacion:
          type: string
          minLength: 3
          maxLength: 500
          example: Se inició revisión documental.
    CerrarSolicitudRequest:
      type: object
      required: [observacionCierre]
      properties:
        observacionCierre:
          type: string
          minLength: 5
          maxLength: 1000
          example: Solicitud atendida, resuelta y notificada al estudiante.
    SolicitudResponse:
      type: object
      required:
        - id
        - descripcion
        - canalOrigen
        - fechaHoraRegistro
        - estado
        - solicitante
        - cerrada
      properties:
        id:
          type: integer
          format: int64
          example: 1
        tipoSolicitud:
          allOf:
            - $ref: '#/components/schemas/TipoSolicitud'
          nullable: true
        descripcion:
          type: string
          example: Solicitud de homologación
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        impactoAcademico:
          allOf:
            - $ref: '#/components/schemas/ImpactoAcademico'
          nullable: true
        fechaLimite:
          type: string
          format: date
          nullable: true
          example: 2026-03-20
        fechaHoraRegistro:
          type: string
          format: date-time
          example: "2026-03-14T10:30:00"
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        prioridad:
          allOf:
            - $ref: '#/components/schemas/Prioridad'
          nullable: true
        justificacionPrioridad:
          type: string
          nullable: true
          example: Fecha límite próxima
        solicitante:
          $ref: '#/components/schemas/UsuarioSimpleRef'
        responsable:
          allOf:
            - $ref: '#/components/schemas/UsuarioSimpleRef'
          nullable: true
        cerrada:
          type: boolean
          example: false
        fechaCierre:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-18T16:22:10"
    PageSolicitudResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, last]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SolicitudResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        last:
          type: boolean
          example: false
...